+++ src/main/java/com/architechz/project/packageservices/models/Product.java
@@ -1,6 +1,7 @@
 package com.architechz.project.packageservices.models;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "products")
@@ -45,7 +46,7 @@ public class Product {
         this.weight = weight;
     }
 
-    public double getVolume() {return volume;}
+    public BigDecimal getVolume() {return volume;}
 
     public void setVolume(double volume) {this.volume = volume;}
